Move set' in the Functional combinators section
I really can't understand why it was in "Simplified state setting"
section. Neigther the type of it nor haddock mentioned something
like "State". Moreover, it's located near the 'set' eliminator
in the source code.

You can argue if it should be in the "Using setters" section, but
I think it's not so generally useful, so it shouldn't be located
in the section of basic functionality